requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==1.0.0
amqp==5.2.0
anyascii==0.3.2
arabic-reshaper==3.0.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.2.0
build==1.0.3
CacheControl==0.13.1
cachetools==5.2.0
captcha==0.5.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
crashtest==0.4.1
crispy==0.7.4
crispy-bootstrap5==2023.10
crispy-forms-gds==0.3.0
cryptography==42.0.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.1.3
django-admin-adminlte @ git+https://github.com/app-generator/django-admin-adminlte.git@567932caf1058d66db620f6a763ce8e202e2bf09
django-adminlte3==0.1.6
django-allauth==0.60.1
django-celery==3.1.17
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-cors-headers==3.13.0
django-crispy-forms==2.1
django-dotenv==1.4.2
django-environ==0.9.0
django-extensions==3.2.3
django-filter==24.3
django-js-asset==3.0.1
django-modelcluster==6.3
django-permissionedforms==0.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
django-taggit==6.1.0
django-treebeard==4.7.1
django-unfold==0.41.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
docutils==0.21.2
draftjs-exporter==5.0.0
dulwich==0.21.7
et_xmlfile==2.0.0
executing==2.0.1
fabio==2023.10.0
fastjsonschema==2.19.1
filelock==3.13.1
filetype==1.2.0
fonttools==4.47.2
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.2
google-api-python-client==2.65.0
google-auth==2.40.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.70.0
gunicorn==21.2.0
h5py==3.10.0
hdf5plugin==4.3.0
httplib2==0.21.0
idna==3.6
imagesize==1.4.1
installer==0.7.0
ipython==8.12.3
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keyring==24.3.0
kiwisolver==1.4.5
kombu==5.4.2
l18n==2021.3
laces==0.1.1
lxml==5.1.0
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.1.3
more-itertools==10.2.0
msgpack==1.0.7
multidict==6.0.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pillow_heif==0.20.0
pipreqs==0.5.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.43
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycurl==7.45.3
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyshorteners==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
pywin32==310
pywin32-ctypes==0.2.2
pyzmq==26.4.0
rapidfuzz==3.6.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roman-numerals-py==3.1.0
rpds-py==0.25.1
rsa==4.9
setuptools==75.5.0
shellingham==1.5.4
silx==2.0.0
six==1.16.0
snowballstemmer==3.0.1
soupsieve==2.6
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.4.4
stack-data==0.6.3
str2bool==1.1
stripe==10.12.0
telepath==0.3.1
tinycss2==1.4.0
tomlkit==0.12.3
tornado==6.5.1
traitlets==5.14.1
trove-classifiers==2024.1.8
twilio==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
wagtail==6.3
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.45.0
whitenoise==6.6.0
Willow==1.9.0
yarg==0.1.9
yarl==1.9.4
